Hoist assets models to global repo
