+++ .eslintrc.json
@@ -6,6 +6,7 @@
     {
       "files": ["*.ts", "*.tsx", "*.js", "*.jsx"],
       "rules": {
+        "@typescript-eslint/prefer-namespace-keyword": "off",
         "@nx/enforce-module-boundaries": [
           "error",
           {
