+++ libs/shared/src/index.ts
@@ -2,6 +2,7 @@ export * from './lib/data';
 export * from './lib/enums';
 export * from './lib/guards';
 export * from './lib/interceptors';
+export * from './lib/models';
 export * from './lib/services';
 export * from './lib/shared.module';
 export * from './lib/state-management/quote-form.actions';
