+++ libs/shared/src/lib/models/index.ts
@@ -1,0 +1,1 @@
+export * from './assets.models';
